requirements.txt
pipecat-ai[daily,openai,cartesia,silero]
pipecatcloud
python-dotenv
loguru